requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.12.0
tiktoken>=0.5.0
httpx>=0.26.0
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
python-dotenv>=1.0.0
alembic>=1.13.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
